LayoutTests/fast/forms/validation-bubble-disappears-when-input-no-longer-visible-expected.txt

Required text input: 
Tests that the HTML form validation bubble disappears when its associated element is no longer visible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS validationBubbleContents.message is "Fill out this field"
PASS validationBubbleContents.message is ""
PASS successfullyParsed is true

TEST COMPLETE
